+++ src/core/similarity_cuda_fetchpair.h
@@ -34,7 +34,7 @@ public:
       ::CUDA::Buffer<float>* expressions,
       int sampleSize,
       ::CUDA::Buffer<int2>* in_index,
-      int minExpression,
+      float minExpression,
       ::CUDA::Buffer<float2>* out_X,
       ::CUDA::Buffer<int>* out_N,
       ::CUDA::Buffer<qint8>* out_labels
